lava spew for orcs that need to cool down
discussed this with sandman when he tried to play an LO in nightmare. being "stuck" at high temperature was very frustrating for him, plus I think this ability is cool and I've wanted to make it since my first days of messing with code on here.

This code is probably terrible, however, and for that I apologize and request assistance when you have time. Here's a rough sketch for now!