quick mode changes
remove movement penalty from quick mode, allowing a player to use quick
mode to rapidly approach an enemy and leave it on, with the only cost
being SP usage per move until the player has a turn to deactivate quick
mode. Also allows players to move around the enemy quickly in a battle.