take Dj up to full health of a normal species
It sounds like having hp be a combination of hp, mp, and sp, and
resulting in having 300 hp early game is way overpowered. But it is
much harder than it appears since each swing of an axe, each square of
running, each spell used, makes significant dents on hp, even if
monsters weren’t hurting the Dj at all. Even on standard mode, this is
proving quite difficult. Hence the change.