Add an option to replace the noise bar with an equipment bar.
The noise bar is great if you are trying to learn which things are noisy.
However, the noise bar is (visually) very noisy itself; it is very
distracting to have it constantly flashing colours and such if you
already have a good sense of which things cause noise.

This commit adds an option "equip_bar", false by default, which replaces
the noise bar with a display of the (glyphs of) items that you currently
have equipped. Usable equipment slots for your character that are currently
unfilled are marked with either "." or " " depending on whether the slot is
currently available (e.g. shield slot is " " if you are wielding a 2-handed
weapon). This bar isn't hugely useful, but at least it provides a quick way
to see how many unfilled equipment slots you have... and a row of shiny
artefacts can be pretty.

Eventually it would be good to have a more customizable HUD where the player
can choose items from a set of many different rows or half-rows, but the
code requires serious cleanup to support something like that.

(This commit also slightly reorders the equipment listing in the % screen:
shield and body armour were swapped, as were left and right ring. It felt
more natural to have weapon and shield slots be adjacent and also have
the first ring slot filled be the first one on the list.)