Regularize qualname and basename for missile brands
Before, both showed prefix brands but not postfix brands, so e.g. a
needle of confusion would have a qualname of "needle" but a curare
needle would have a qualname of "curare-tipped needle".  With this
change qualname shows both prefix and postfix brands, and basename shows
neither.  I base this on the behavior of other items, such as scrolls
and potions.  This change may fully resolve 3062.