Fix conjure flame on self messaging (Avigdore)
Instead of using the 'suicidal' message, use the message for
non-damaging spells that can't target self (in a very indirect way).
This is a quick fix, and arguably `direction_chooser::targets_enemies`,
which doesn't make a lot of sense right now, should be generalized to
handle spells that have to target empty spaces. Right now this concept
is implemented as a special case for particular spells, rather than a
SPFLAG_ or TARG_ value. Setting SPFLAG_NEUTRAL gets the targeter to be
TARG_ANY instead of TARG_HOSTILE, which then is interpreted as something
that doesn't target enemies.