Add a (c++) validate function for mutations, including Ds muts
This reveals that many Ds saves with non-trivial non-innate mutations
are broken, because bookkeeping for the case where a non-permanent
mutation becomes permanent on levelup is not done correctly.  Fix coming
in a future commit.

The intent is to eventually call this validation function on load, but
doing this without the Ds fix will create a whole bunch of save compat
issues.