Scale cell widths/heights by devicePixelRatio (wheals)
Many browsers, though not all, change devicePixelRatio when the user
zooms in or out.  Since we scale the canvas by this (e.g. in order to
handle high DPI devices properly[1]), zooming was messing up
calculations of mouse position for purposes of selecting cells.  This
commit scales cell dimensions appropriately.

I've tested this on firefox, chrome, & safari on OS X.  It fixes
problems with mouse cursor position and zooming on firefox/chrome, and
doesn't break safari (which already worked).  It needs testing on other
platforms, in IE, and on high DPI devices (such as retina displays).

[1] refs for scaling reasons: 75d9865c86980b7, 0475afac5cde