round out the book of power
this rare and expensive book has every element in it except for ice. Also, it keeps shrinking - note the learndb compared to this!

Bolt of cold would be a better choice, because Servitor can actually cast it. Maybe Fcloud as well, but that's already in multiple other books. There's always icicle, too.