Implement training target backend
This commit implements the basic backend machinery for skill training
targets, as well as a UI for viewing them in the m screen as another
mode. Beyond the lua bindings, there is no UI for editing them in this
commit, that will come in a future commit.

Each skill can be associated with a unique training target, implemented
as a 10-scaled skill level value.  When the target is reached or passed,
the training target is reset to 0 (which means 'off'), and training for
that skill is turned off.  Targets are checked while training, while
setting skill levels, and on load/skill fixup.  Because of the way skill
training works, targets aren't checked exactly, so a large burst of
experience (e.g.  killing cerebov, drinking a potion of xp) can
overshoot targets, but in practice it gets pretty close.  Targets are
checked against skill factoring in cross-training and ash bonuses, but
not draining. With this commit, training targets can be set and accessed
via a pair of lua bindings, you.get_training_target and
you.set_training_target.