Prevent casting a spell with 100% failure chance (aidanh)
If there's never an advantage for a player to try to cast a spell with
100% failure, why do we even allow it as an option?

If there's really some edge case advantage for a player to try to cast a
spell with 100% failure, why would we continue to encourage that?

Seems like a win/win to just prevent it from being cast in the first
place.

(cherry picked from commit 505e61d45a1a5d68e5dd89f6d3f3f2f77fa49a2d)