Add UI for skill target editing for console/local tiles
This bolts together TextItems and line_readers to make a (very) basic
text editing widget (which previously only existed in the form of
line_reader in the message pane), and uses it to implement editing for
skill targets in the m screen. When viewing skill targets, you can
select a target to edit, and the editing happens in place in that part
of the skill menu.

As part of this process, I attempted to make line_reader very slightly
more sane in a number of ways; this commit also includes a local tiles
subclass that uses a FontBuffer (for arbitrary positioning of the
editing widget, rather than grid positioning).

This commit doesn't cover webtiles, and editing only sort of works there
(it works basically like on console, but the cursor doesn't show).  The
next step is to generalize the input box technique currently limited
only to the message pane.