Remove some doubles, and fix a wrong assumption about shields
Weapon mindelay is generally in integers, so in parts of the description
code I was discarding fractional parts. However, shields can have
non-integer training targets, e.g. bucklers for a formicid have the
penalty removed at 2.4. In an earlier commit to try to streamline the
description code, I switched part of the shield calculation to integers
incorrectly.  This version now uses 10-scaled ints for everything and
does not assume that any item-based target will be integer.