Push items out of the way when a door is closed, if possible
By popular demand, this version pushes items out of the way when closing
a door, but refuses to close a door if there are items that can't be
moved anywhere. (This can come up with deep water, lava, but also when
something like door_vault with a gateway places against a wall.) It
should be impossible to get items under doors on this version.