Add prompt for wall jump
Adds a GUI prompt before performing a wall jump. It is now necessary to
walk against a wall twice to actually jump (although it does not need to
be the same segment of wall, in case you change your opinion halfway).

There is a RC file option (wall_jump_prompt) to disable this for players
who are used to the old behaviour.

Hopefully this will significantly reduce accidental wall jumps without
forcing the user to detour through the ability menu, and since wall jump
spam is not a thing anymore, the extra keypress is not an issue.

Resolves #658