Refactor some code into get_tiles_for_item()
The code that adds the background tile assumes that the item is on the
same level, which is not always the case, and generally makes this
pretty buggy. But a level_excursion would likely be too heavyweight.
Blegh.

(cherry picked from commit 285f75aafb25dd1112bf0839dd8e057fbbe0e764)