Rewrite text input suppression to use different hacks
On most(?) windows and linux setups, numpad keys with numlock on send
both a keydown event and a textinput event with the corresponding
number; with numlock off only the keydown event is sent. On OS X there
is no numlock, but this dual behavior happens always. Crawl is supposed
to treat the numpad uniformly regardless of numlock state, so it needs
to ignore the double press. For legacy reasons (I think) it also needs
to correctly handle just the number input. In the numlock on case (and
on OS X), therefore, one or the other has to be suppressed.

The previous approach relied on starting and stopping the text input
system clearing any upcoming SDL_TEXTINPUT events. On Windows, updating
SDL to 2.0.7 has broken this assumption, and this toggle only sometimes
works as intended. I couldn't figure out a reasonable way to restore the
behavior of this toggle, so here is a new approach.

This commit tries to do things without relying on any particular SDL
code; if a textinput event immediately follows a matching keydown event,
the textinput event is ignored. This of course assumes that these events
are adjacent, but in my testing on OS X and windows this has always been
true so far. This patch will need some more testing.

(I suspect the relevant SDL commit is this one:
https://github.com/spurious/SDL-mirror/commit/ff8a1efc51a6ad7de6379cde7047d135ffd7e15e
because it changes the timing of textinput events, which were previously
guaranteed to be near the keydown; but I haven't tested this
hypothesis.)