Fix ?/I showing descriptions for wrong items
There were several problems with the old implementation.
get_item_by_name() did fuzzy matching, instead of exact matching, and
only a handful of item types were checked in _describe_item(). In
addition, a partial match against item types checked first would be used
immediately, without checking for a better match of a different type.

To fix this, get_item_by_exact_name() is added. The name given to it is
compared against the dbname of all possible items, and only an exact
match is considered a success.