Proof of concept implementation of server-side tileschat mute
This implements per-process muting for tileschat in the webserver. There
are three commands, "/mute", "/unmute", and "/mutelist"; only the player
can use any of these commands. Right now this doesn't save at all, so
would need to be reconstructed on every save/reload. It's just a first
pass / thought experiment in how better tileschat controls could be
implemented, so more work would be needed before merging. There's no
indication to anyone except the player who is muted, which might be good
to have.

I'm not sure this is the best way to do it, especially for a quick fix
for spamming. In ##crawl-dev gammafunk and I discussed an alternative
that would be client-side, per-user (instead of for the whole channel),
and saved in a cookie, which could be better, and wouldn't require any
server updates. (It might be possible to adapt the cookie saving idea to
the current commit.)

(cherry picked from commit 186421b65b3d3b8607a3b959be277529cdc0d52d)