Allow searching for spells in the memorise menu.
The search matches either on spell names or spell types (so you can see
all the translocation spells at a glance, for instance).

The extra newline was removed from menu.js since it seems the browser
adds its own newline (tested this with the \ menu and it works the same
as before).

Added a bunch of other menu refactoring stuff I didn't merge earlier, apologies if this breaks anything.