Add HTML paneset construct
This is the HTML equivalent of the UISwitcher widget: a set of widgets
(divs) in the same position, only one of which is visible at any one
time, that all maintain the same size.

It's implemented as a 1x1 grid, but that might need to be changed for
browser compatibility reasons.

(cherry picked from commit f7378a7a1895bbcca1501e3a7c90e64b047e53ee)