Make statue descriptions accurate about no-regen (#11448)
This commit gives statues the M_NO_REGEN flag, which will show up in
their description as "It cannot regenerate".

This wasn't there before because statues cannot regenerate by any means,
as seen in the check in monster::heal, whereas other monsters with this
flag will still regenerate if given the regeneration status. The check
in monster::heal is still checked with this flag on, however, so there
is no reason this should change any behaviour.