Fix the octopode + P interaction when wearing many rings (aubilenon)
When an octopode is wearing all the rings, and that number is more lines
than there are lines in the message window, using P and selecting a ring
forced an extremely awkward prompt that required a more, and then a ? if
you want to see all the rings at once. This change just has the prompt
use the menu from the beginning if there are too many rings to fit in
the message window at its current height (the normal situation for
webtiles).

Annoyingly, there is no easy way to prompt only for rings, so this
change shows all jewellery (for now).

(cherry picked from commit 535178c5c5b0bba1c9da54fed99a4b18539abe5d)