Ensure that close_all_menus syncs with webtiles clients
This is one possible explanation for 11387, that somehow the javascript
menu_stack gets out of sync with m_menu_stack on the client side. I
haven't managed to reproduce this bug at all, but the two stacks getting
out of sync via an extra open_menu immediately following init_menus
would lead to the observed behavior, and explain why further menus act
as submenus. This change ensures that the javascript stack is entirely
empty when client-side close_all_menus is called, which is triggered on
redraw. If this fix doesn't work, back to the drawing board.