Clean up bones file debugging messaging
This centralizes all the complicated #define conditioning in one place,
so it's a lot easier to figure out what's going on, and hopefully this
will minimize accidental compiler warnings. This relies on a fairly
hacky unwind in one place, though.

This is all a bit complicated because one wants these messages to show
in wizmode on a server, which won't ever have a debug build.

(cherry picked from commit 3ff10968076cbfd2dca25221d3e1ce2ca842ad22)