Fix abyss_short_run.rc (aidanh)
This is doubly-broken: Since 1f733d267d2fff9e920c items of flying start
flight when worn, and at some point (not sure when) the default key for
starting flight seesm to have become `g` anyways. So this test got a
silent error on sendkeys('ah'), and failed to ever enter the abyss.