Have display of the acrobat ev boost in tiles match console
Neither the number, nor the color, were showing properly in webtiles,
where by properly I mean showing with the timing that they do in
console. (As discussed in the dev channel in the past, it isn't entirely
clear what the best way to show this is, but I went with the existing
behavior.) To get the number to show, the LAST_MOVE_... property needed
to be reset slightly later in main.cc, because it was getting cleared
before the webtiles redraw (and probably regular tiles too, though I
didn't test it); this led to it only appearing under strange
circumstances such as resting with 5. To get the color to render, it
just needed some explicit casing to get the timing right. As of this
commit, the behavior should be the same in tiles/console.

I also slightly generalized / unified some of the logic for checking
whether to show the acrobat boost.

(cherry picked from commit 8519d868e99a0bfa83bc6c418198c0275198a07a)