A "magic classroom" special room for Depths (Skrybe, #11384)
A Depths special room that places spell-casting monsters with small
chances to place one of a few spell-using uniques that normally appear
in Depths.

This only non-formatting changes from the original submission are to add
Boris to the unique list and to tweak the least dangerous case that can
happen on Depths:1-2 based on the result of a random roll. For that
case, place a tengu reaver as the alternate for a lich instead of having
small chances for said uniques. These uniques normally spawn through
their vaults on Depths:3-5 and are certainly more comparable to fighting
an ancient lich that places in the worst case in this special room.
Therefore give them weight only in that worst case monster list.

(cherry picked from commit c4eb3cb44a9748055c5479521d3bc817dd573e8b)