Fix shopping list crashes from multiple items with the same name
Mainly happening when trying to remove spellbooks from the shopping
list.  It was possible to get multiple items with the same name on the
shopping list, which would then (depending on iterator order) lead to
crashes if they were removed. All of the find/delete code behaved
unreliably in this case, because it worked by getting the first index by
name, which may not actually match up to the `thing` that was passed in,
and so the removal order didn't match up with the iterator order.  This
change generalizes these bits of the shopping list code to not assume
that there will be single items associated with any name-level pair.
This fix is more of a patch than a complete fix, because:

1. The level_pos checks in this code seem to be still unreliable in ways
   I haven't fully tracked down yet.
2. Reading the code, it was supposed to be impossible to have multiple
   items with the same name, but the checks for this are hopelessly
   broken (as far as I can tell), and probably undesireable anyways.
3. It seems to me that associating shopping list items with level ids,
   rather than particular shops, is just a fundamental design flaw that
   makes it impossible to uniquely associate shopping list entries with
   actual particular items in the game. This will need more of a rewrite
   to fix, though.

(cherry picked from commit 10d5862e6fc6b375f1b71f8f77347b7f5b30896d)