Fix webtiles status artifacts due to custom status light code (11488)
When a duration was immediately followed in duration_type by a duration
that has some custom webtiles code in tileweb.cc:_update_statuses, the
status_info wouldn't be fully reset for the custom status; I'm not sure
if this had any noticeable effects beyond the acrobat case, but this
would've bitten someone eventually (it's more or less by accident that
divine shield follows divine stamina, so one couldn't have both at the
same time).  For acrobat in particular, 8519d868e99a introduced a check
that was missing a continue, meaning that recite cooldown (the next
duration in the enum) would show up twice if acrobat was not active.

This also replaces a janky static reset function for status_info with a
proper constructor that actually guarantees the int is initialized, and
switches unnecessary pointer uses of status_info to references.

(cherry picked from commit 8c8ef71062e55278bb8735b2e84cec8f08405607)