Replace remaining uses of `*map_find` with explicit ASSERTs
`*map_find` is highly dubious, because it assumes that the relevant map
is complete with respect to some enum (usually under some specific
assumptions), and none of the uses were plausibly cases where this is
just going to automatically be fine in the long run.  This change
converts the checking of that assumption to an explicit ASSERT, rather
than a segfault.

(cherry picked from commit 19b682e6188f74e086018dec68c799ce639e5a16)