Revise the mace of Variability
Having a weapon that was occasionally useless didn't make for a good
item. In the tradition of other unrands that are "this brand but more"
(Arc Blade, Firestarter, Frostbite, Leech), this changes the mace to
have a chance of casting chain of chaos on hit and have a fixed plus. As
flavourful as the plus swing is the player is encouraged to wait for the
random walk to go up before fighting.

Committer's Note: The chain of chaos damage and set of effects as well
as the enchantment for this weapon are modified in a subsequent commit.

(cherry picked from commit 8c3e449617027ffac50e618f75cb7e63d1cd40f5)