Add elemental colour support to the equip bar
This commit modifies the equip bar to process elemental colours (coming
from unrands) for equipment glyphs in the equip bar.

By default, the colours are not animated, since a goal of the equip bar
was to lower visual noise. The option "animate_equip_bar" is added to
toggle this behavior, as the other goal is to show of your cool stuff.