Add dispersal traps
Dispersal traps are permanent, and when triggered cause everything in
sight of the trap to blink, including the triggerer.

The purpose of dispersal traps is to provide an alternative for known
teleport traps that is more interesting tactical terrain. The effect
destroys a player's positioning, so if it is good the player will want
to prevent monsters from stepping on the trap. On the flip side, it can
provide a positional rearrangement in a bad situation, though preserving
a good result may be challenging since the trap is permanent.