Adjust the placement of Ilsuiw (alexjurkiewicz)
Currently Ilsuiw has a 100% chance of placing on Shoals:4 if she didn't
place on Shoals:3. Fighting her will be a more memorable (some would say
'unique') experience if doesn't happen every time on a specific level.
This commit introduces a 1/5 chance that her Shoals:4 guardian vault
places. It also sets the range of her normal unique vault to Shoals:2-3,
both so that she can place on Shoals:2 and so that this vault does not
give an additional chance of placing her on Shoals:4. Finally, we give
her normal unique vault the typical weight of 50 we use for
branch-specific uniques. Now her chances of placement on Shoals:2-3 are
more similar to that of placing on Shoals:4.

After this commit, mapstat shows her chance of placing overall in shoals
is between 40 and 50% (more objstat iterations needed to be sure), with
a slightly higher 20-25% chance of placing on Shoals:4 compared to about
10-15% chance for both Shoals:2 and Shoals:3. This is similar to
placement of other branch-specific uniques like Polyphemus, Arachne, and
Vashnia. Closes #774.

(cherry picked from commit bd89b2e156b50a805e4d5adcda5fc102a0f27db4)