Switch from LuaDoc to LDoc
LuaDoc never had C++ support, and is now listed as obsolete, directing
users to switch to LDoc, which has C++ support (if a few quirks of its
own).

Adds a make target api-dev to generate a developer's api documentation.
This includes internal functions and hooks that the user should not
touch directly.