Make socket write retries more aggressive
Under some circumstances, these writes may block while the webtiles
server processes data -- something that happens very quickly, but not as
quickly as the send loop in this code. For example, if the socket buffer
size is <=4096 (the default on OS X, but not linux), trying to send
larger batches of data (such as a map on load or level change) will
cause ENOBUFS until the server has had a chance to receive the previous
fragment. The previous sleep timing was effectively causing a 500ms
sleep per 2048 byte fragment[1], and so since a map can take on the
order of 200kb, this was imposing substantial slowdowns on OS X
webtiles.

This fix first tries some 2ms sleeps, and then backs off to 500ms. 2ms
is still more than enough for the webtiles server to do its thing in my
testing; we'll see if it causes in problems on live servers (I have no
idea how often this case is triggered outside of OS X). I may still
accompany this with a change to the default socket size.

1 I'm not sure why it's not every two fragments, but there might have
been a few initial bytes in the buffer in real-world scenarios, or some
overhead not directly accounted for.

(cherry picked from commit d5849c758ab4318eccdb0db425ea5cd5529cf629)