Don't render UI layers under direction chooser (stoneychips, #11525)
This fixes a webtiles bug where evoking items from their inventory item
description would lead to the inventory menu obscuring the direction
chooser. This also fixes a secondary bug on local tiles where evoking
items from their inventory item description and then pressing v to
describe a cell would cause the inventory menu to appear under the
description popup.

Briefly, there are two main render loops for local tiles: the directn
loop (which does not render UI popups), and the new UI loop (which does
render the background direction chooser view). Effectively, either all
popups or no popups are rendered, depending on which render loop is
running on the top of the stack (they can be nested). For xv, that's the
UI loop on top of the directn loop; for inv -> item -> evoke, that's the
directn loop on top of the UI loop.

This approach doesn't work on webtiles because it uses HTML to render
the UI, and so the render loops are effectively "running in parallel".
It was buggy on local tiles when describing cells after evoking an item
from the inventory. The fix is to add "cutoff points" which prevent the
rendering of any UI layers already pushed. Local UI and tileweb have
separate UI stacks, and so have their own cutoff point API, in case
non-UI code modifies the tileweb UI stack in the future.

The addition of the cutoff point to show_map() is currently unnecessary,
since there is currently no way to show the map from a popup view, but
this might change in the future.

(cherry picked from commit 058c00fc3bc5f744541a21adff4f14aa3370a387)