Update popup UI skin
This commit removes the image-based background for UI popups, and
replaces it with a border using solid colours. This simplifies the
implementation on webtiles significantly. The border colour is changed
to match the scrollbar colour.

This commit also unifies how popups are aligned within the screen for
webtiles and local tiles, with some changes for both. When a game
has started, popups on local tiles are now aligned top-centre and
offset vertically according to their nesting depth, matching the popup
alignment on webtiles. When a game is not in progress, however, popups
on both platforms are aligned to the centre, and there is no vertical
offset to indicate nesting level.

In future, it'd be nice to a) deduplicate all the different places where
the centredness of the popup is sent to webtiles, and b) query whether
the popup is centred from the popup widget itself, allowing for finer
control.

(cherry picked from commit 5fbd67152cc9c3d359f31a17de24f9d78c2f793e)