Update webtiles UI scrollbar skin and positioning
Moves the scrollbar outside the scrolled content and against the right
popup margin, similar to the local tiles positioning. This is done by
adding a -20px margin on the outer scroller container, which matches the
20px padding of the popup boxes. This negative margin is offset by a
padding set in scroller.js; necessary, because simplebar.js changes the
padding set based on the current browser.

In addition to the positioning changes, we also style the scrollbar
handle with the new skin. We also give it wide-but-invisible borders, so
that it's easier to click on and drag, and a nice cursor icon.

This is unfortunately fairly brittle, and there's some browser
differences which make this code more complicated than it should be.
Webkit-based browsers do not seem to force inline-grid children to
shrink to fit a smaller container. To hack around this, we set the
height of all paneset children with every browser resize event. This is
mostly an issue only on the ^ UI.

There are a few minor bugs. The first is that vertical scrollbars do not
appear immediately when a scrollable popup opens, even when content can
be scrolled. This seems to be due to building the scroller UI with
simplebar.js before the popup is added to the main DOM hierarchy.
There are also some minor glitches when changing the zoom level while the
^ UI is opened.

(cherry picked from commit 9cf2c4cdec58069b18da7ec106a05ec56f83cc9d)