Adjust a Lair entry vault (minmay)
This vault was designed around the guardian being able to fly, but when
fire drakes were removed, the hellhounds that replaced it tends to just
get stuck. This commit substitutes the hell hounds with a fire dragon
and gives this entry the lowest possible weight. A fire dragon is not
too bad an OOD for D:8-11, being equivalent to a 8 or 9 depending on the
Lair entrance depth. This will be a rare, nasty guardian, and of course
players wanting to enter Lair but who are unable to kill the dragon can
employ various means to lure it away from the entrance.

(cherry picked from commit 9a2f62deabe5e2ba6b4d9cc5ea53ed810698aa1a)
(cherry picked from commit 5fd64747bb5cd929afb710d3a80f22bf6fa634e9)