Convert Crypt to single level
Crypt is a bit of a nothing-branch right now. It's mostly known for
being an otherwise safe place to burn of god wrath for your 15-rune
TSO/Makh god switch. There is decent end-game loot, but it's outshone by
Elf, and since the only sub-branch is Tomb, it's rare to visit here
outside a 15-rune game. The theme of the branch also overlaps
uncomfortably with Tomb (mummies, undead) and Depths (vampires).

This series of patches will make several changes, but the first one is
to convert the branch to single level. There are some great Crypt ending
vaults, but the layout engine for other levels is nothing special. A
single level gives more room for a smaller, tighter experience in terms
of monster set size and gimmicks, so they can be quite different but not
overstay their welcome.