Fix webtiles scrollbar visibility and zoom issues
The purpose of these modifications is to make the scrollbar 'stick out'
from the side of the scroller window, so that it is flush with the edge
of the popup. Unfortunately, using a transform is not possible here due
to clipping.

This commit resolves some issues reported by users where the scrollbar
would not be immediately visible, and where the browser scrollbar would
reappear if the zoom level was changed.