Refactor LOS attack spell code
This code previously had a lot of duplicated logic between functions
used for the two spells that cast it, Ozo's Refrigeration and Drain
Life. Instead of using function pointers to different functions doing
similar things, use common functions where possible.

Make zap entries for Ozo's and Drain Life to store details like the
damage and color (used for flash effect). This way these details can be
stored in the beam already used to store attack details.

This commit changes the damage formula of Drain Life, which was
previously:

3 + random2(7) + random2(pow)

to be:

2d(4 + pow/2)

which has a similar average as well as min/max damage compared to the
old formula.

Bcadren: Fixed a conflict caused by me saving a future fix to a bug caused by this patch
in case I needed it later. I guess when I get to that bug fix in the tree I'll already have it.