Revamp the Singing Sword
Currently this unrand is worse than a +7 vorpal double sword, since it
has no brand and its "sonic damage" effect is weak. This sonic damage is
also linked to having the sword wielded instead being linked to
successful attacks. This commit rebalances the sword to be more
comparable to an enchanted double sword of speed at high tension. It
also makes the sonic damage effect to trigger only on successful melee
attacks (including ripostes). The sword now casts a "Sonic Wave" spell
that uses existing LOS attack spell code, instead of duplicating some of
that logic, and has a visual flash when the effect triggers.
Additionally the sword has a vorpal brand, instead of being brandless,
and has an inscription and updated description indicating its spell
effect.

The sonic wave damage is simplified from the previous multiplicative die
to use a 2d(5 + pow / 10) roll with the chance for the effect lowered to
50%. When the effect triggers at high tension, the sword has a 1/10
chance to "scream", as it did before this commit, but this now gives 50%
more spellpower instead of double damage, since the sonic damage is much
higher now.

Finally, the sword's noise has been set so that it's the same chance and
20 noise value as the *Noise property at low tension, with that noise
increasing by 10 at high tension and by 10 again if the sword "screams".