Fix a crash for tracers of LOS attack spells
Introduced in fb95405 where an average damage value was removed. This is
still needed for tracer calculations to get a scaling factor for friend
info based on what proportion of damage the target would resist.

This commit also refactors some of this code to remove redundant
variables and better use an input argument pointer to hold the total
damage.