Don't prevent the ability menu from opening when berserk or confused
Previously, three things happened depending on the player's statuses:
  - if the player was berserk, the ability menu would never open;
  - if the player was confused, the ability menu would only open if the
    player had one of the two conf_ok abilities (Ely's Purification and
    Ru's Draw Out Power), and if it did open, only conf_ok abilities
    would be shown.
  - if the player had any other restrictive statuses, the ability menu
    would open, and, in most cases, abilities affected by the
    restrictive statuses would be greyed out.

It was decided that having berserk or confusion act differently to other
statuses, for example silence, was inconsistent and unnecessary. Hence,
we allow the ability menu to be opened even when berserk or confused,
where unusable abilities will be greyed out.