Fix two iterators becoming invalidated during iteration
Both of these loops, used in constructing shoals levels, can add
something to the back of the supporting vector during iteration, which
(per the spec, though who knows in practice) will immediately invalidate
an iterator for a vector. This change is a partial revert of
a7d9996df05.

Found via #854.