Revise the robe of Misfortune (gammafunk)
The robe of "particularly bad identification trap" was not an
interesting thing. This keeps the theme of an unlucky robe which
frequently leads to unfortunate confluences of events, but now lets some
of that bad luck fall on monsters too. Specifically, the random
appearance is dropped and the robe becomes: +5 with {*Corrode *Contam
*Drain Harm *Slow EV+5}.