Rework the Sceptre of Torment
This torment effect of this weapon makes it something only undead
players want to use, and even then it's not a very good weapon due to
having such low direct damage. This commit makes the torment effect not
apply to the wielder, applying it on 20% of successful hits instead of
50%, changes the base type to be an eveningstar, and gives it the pain
brand. It was previously an unrand one-handed staff, but we already have
one of these, the Elemental Staff, and had no unrand eveningstar with
the removal of the Mace of Brilliance. Now the sceptre will be a usable
weapon that outclasses most branded and highly enchanted weapons for its
type.

It might be nice to update the tile to have spikes, so it more resembles
something one could conceivably call an eveningstar.