Allow setting attitude of cloned monsters (damerell, #11171) (#608)
Matches the attitude of a cloned monster to the cloner. This also
removes the (currently never-used) option of setting the position of the 
cloned monster on call.

For _mons_summon_monster_illusion:
By passing the cloned monster's attitude to clone_mons, we don't have to
do all of the gymnastics that previously had to be done to properly set
the cloned monster attitude and not adversely affect arena bookkeeping.

for _phantom_mirror, we can now simply set ATT_FRIENDLY in the
clone_mons call directly.