Add a lighthouse-themed bazaar vault
Adds the vault bazaar_kenran_lighthouse. It features a circular hut
built on a small island, make to resemble a lighthouse in the ocean.

The shoreline, as well as the transitions from shallow to deep water,
and from deep water to open sea, respectively, vary slightly in each
generated map.

At the moment, the centre of the lighthouse contains a trapped
ophan (as to provide a source of light). Of course this is rather
gimmicky, but I found it fun. It can easily be replaced by another
shop or decorative interior though, should monster generation in
this case be frowned upon.