Improve (?) filtered_vector_select
This now takes the `filter` argument in a format that should be
guaranteed to cast to bool (for purposes of #871), guaranteed to handle
a nullptr argument, and is also just more explicit about its type.
Unfortunately this prevents at least my compiler from deducing M
correctly, but such is life.