Fixes, refactoring, and message cleanups for static discharge
The safe_discharge() function for the attack prompt warnings of static
discharge didn't properly consider the player as an arc location. This
meant that casting it didn't give a warning in some cases where an ally
could be damaged.

Also refactor the main cast_discharge() spellcasting function to allow
non-players to call it, then have monsters use this when casting the
spell. Previously the monster casting code partially duplicated this
function.

Finally, simplify the messaging a bit to remove a separate message about
lightning arcs, adding that to the main damage message instead. The
message spam from this spell is pretty high and could probably be
improved to only print wound messages after all arcing is completed.