Fix a crash for LOS attack spells (MeekVeins, amalloy)
It's possible for a monster to die after applying damage flavour effects
but before the direct damage is applied. This invalidates the monster
pointer needed for god conducts, and was causing crashes with e.g. Drain
Life. This commit adds a necessary check for the monster being alive
before trying to apply conducts.

Additionally Drain Life shouldn't actually the flavour effect of
negative energy, and this will be fixed in a subsequent commit.