A Gauntlet portal to replace Labyrinths: renames and code removal
Labyrinths disable autoexplore and force the player to manually solve a
maze, something not related to the kind of gameplay crawl tries to
provide.  There are spoilery aspects to navigating the maze efficiently
and to using teleport in the portal. Once these aspects are known to an
experienced player, they solve the maze with very little thought and
often with great annoyance at having to play this mini-game unrelated to
either tactical combat or strategy. The most interesting aspect of
current labs is the fight with a single powerful monster, the minotaur.

This commit is the first in a series to redesign the portal to be a
gauntlet of arena fights where the player chooses their path. The new
portal emphasizes this decision, giving the player information about the
risk they'll face and rewards they'll get. The portal is similar to
Ziggurats in terms of having progressive rooms with thematic monster
sets, but adds a distinct choice aspect.

Each of the possible paths takes the player through a series of
randomized arena subvaults that have one of two difficulty tiers. Each
arena subvault will having a small group of thematic monsters and a
small amount of loot. All possible paths that could be taken will lead
to the same exit subvault that has boss fight with the minotaur and a
loot pile on the exit stairs.

The number of arena subvaults in each path will be small, usually two
but it could be three or four on some maps. The vaults will will be
accessed via transporters that prevent backtracking by simply walking
out of the vault, and the arenas will not have diggable walls.
Allowing or disallowing teleport is one aspect under consideration for
this portal, but currently teleports are allowed.

Before choosing a path, the player will be able examine all arenas in
each path to see available monsters and loot. At least one early exit
will be provided at the beginning of the portal before any path is
taken.

This portal isn't designed as a maze and doesn't fit the traditional
idea of a labyrinth, so the new portal will be called a Gauntlet
instead. This commit renames file and data to take on the new name.
Since we're re-using existing enum values, save compat won't be
necessary. Games saved in a Labyrinth will load and the lab will function
normally, save that lab shifting will not occur.

This commit also removes the C++/lua code and vaults that won't be
directly used by the new portal Some of the minotaur exit vaults will
show up in the new portal with modifications, but we remove these for
now. Many of the removed float vaults and exit vaults will show up in
subsequent commits, reworked for the new portal.