Gauntlet: A new portal that replaces Labyrinths
Labyrinths disable autoexplore and force the player to manually solve a
maze, something not related to the kind of gameplay crawl tries to
provide.  There are spoilery aspects to navigating the maze efficiently
and to using teleport in the portal. Once these aspects are known to an
experienced player, they solve the maze with very little thought and
often with great annoyance at having to play this mini-game unrelated to
either tactical combat or strategy. The most interesting aspect of
current Labs is the fight with a single powerful monster, the minotaur.

This commit is the first in a series to implement a new Gauntlet portal
where the player chooses a path through several combat arenas. The
player can see what monsters they'll face and what rewards they'll get
before entering each arena. The portal is similar to Ziggurats in terms
of having progressive rooms with thematic monster sets, but adds a
distinct choice aspect.

Each of the possible paths takes the player through a series of
randomized arena subvaults that have one of two difficulty tiers. Each
arena subvault will having a small group of thematic monsters and a
small amount of loot. All possible paths that could be taken will lead
to the same exit subvault that has boss fight with the minotaur and a
loot pile on the exit stairs. At least one early exit will be provided
at the beginning of the portal before any path is taken, with the only
other exit being guarded by the minotaur.

The number of arena subvaults the player completes in each path will be
small, usually two but it could be three or four on some maps. The
vaults will be accessed via transporters that prevent backtracking by
simply walking out of the vault, and the arenas will not have diggable
walls. All forms of teleportation will be disabled in this portal in a
subsequent commit, but other translocations will work normally.

This portal isn't designed as a maze and doesn't fit the traditional
idea of a Labyrinth, so the new portal will be called a Gauntlet
instead. This commit adds basic data for the new portal and renames any
files being repurposed from the old Labyrinth portal.

This commit also removes the C++/lua code and vaults that won't be
directly used by the new portal. Some of the Lab float and exit vaults
will show up in the new portal with modifications, but we remove these
for now.

For any games saved with an active Lab entrance, that feature is
replaced with an entrance to a Gauntlet. Games saved inside a Labyrinth
will load and the Lab will function normally, except that Lab shifting
won't occur. Since the player is always guaranteed a viable path to
the exit after the shift, the Lab will still be solvable.