Properly notify non-users of Gourmand (11692)
Previously, the wear message for gourmand was the same regardless of
whether it would work; change it to reflect the various reasons why it
might not work. Carnivorous/herbivorous mutations cover most species.
`foodless` covers mummies and the extra case of lich transformations,
for added lore. Vampires don't seem to fit into either of these cases,
so they are checked explicitly.