Limit extra Ru ability info to followers (amalloy)
The Ru sacrifice ability description contains code for showing sacrifice
significance and piety gain, which crashed if the player was not
following Ru for some abilities. This commit simply hides that line
of text, as well as mutation-specific text for sacrifices that add
mutations, and re-enables using the ability description view for
?/A, which is currently the only way for non-followers to view the
description.