Remove a vault that breaks autoexplore (minmay)
Removed by request of the vault author. The vault
minmay_item_behind_shaft generates a corridor with a shaft at the end in
front of an item. If you recognize this vault you'll dig into it,
apport, use allies to trigger the hidden shaft, or just return when the
level below is cleared. If you don't recognize it, you'll autoexplore
into it, always triggering the shaft. To some extent all vaults that
place hidden traps are spoilery, but this is an egregious case. Consider
this elf flogged.