Validate file pointer when loading help parts
This might be better to handle without a crash at all, but right now
there is a possible segfault in fgets if something is wrong with this
file, so check it explicitly. Possibly these files should be explicitly
validated on game load in addition (I think only their containing folder
is right now).