Rest the Ancestor
Add a delay interrupt that executes when the ancestor heals to full HP.
This is facilitated by a new check for ancestor HP being full; this is
always true if the ancestor doesn't exist or is out of sight (to avoid
information leaks).

Introduce a new boolean option rest_wait_ancestor to toggle whether resting
should always wait until the ancestor is healed. This interacts with
rest_wait_both according to the following table:

| rest_wait_both | rest_wait_ancestor | expected behavior              |
| -------------- | ------------------ | ------------------------------ |
| True           | True               | Rest until (HP full and
                                        MP full and Ancestor full)     |
| False          | True               | Rest until (HP full or MP full)
                                        and (Ancestor full)            |
| True           | False              | Rest until (HP full and
                                        MP full) or (Ancestor full)    |
| False          | False              | Rest until (HP full or MP full
                                        or Ancestor full)              |

[ Committer's notes: Closes #900. Tweaked some conditionals, adding some
  missing conditions or simplifying. Added some comments. Tweaked the
  wording of the option. Adjusted the interrupt to not leak information
  about the ancestor when it is out of LOS. ]