Display pacification chance in Heal Other targeter
With the changes to pacification to depend only on numbers visible to
the player, it is now not an information link to display pacification
chance in the pacification targeter.

This also changes pacification to abort when used on monsters which are
truly impossible to pacify (as opposed to those whose pacification
chance rounds to 0% in the display).