Nerf player-to-monster confusion
Confused monsters now always try to set a target. One chance in three a
monster is able to move toward or attack its target; though spells, most
abilities, evokables, etc. are still forbidden. Confused monsters are
still stabbable. If this proves too subtle a nerf this can be upped to a
coinflip.

The reasons for doing this are twofold. One, Confusion as an effect is
available to players at level 1 or 3 in spells, the latter in a starting
book. Confusion was effectively a temporary instant kill, which is an
extreme effect to provide at such a low level, even with checks. Two,
the temporary instant kill was annoying to make permanent because the
monster would wander away while the player attacked it.

This commit also cleans up some of the mon-act code, specifically
removing some dead confusion handling code from the movement handler
that was not executing before this patch.