Gauntlet exit subvaults
These vaults place a final portal exit, the minotaur, and the loot pile.
All are adaptations of the old lab portal exit vaults to work with the
current layout. The changes I've made are to rotate all vaults to the
same north-to-south orientation, make them subvaults, and to give the
smaller vaults surrounding area to help isolate the minotaur's lair.
This prevents noise from drawing out the minotaur to "camp" the player
near the final transporter landing site.