Disable player teleportation in Gauntlets
This portal's layouts have the player choose a one-way path through
arenas with monsters and loot. Teleportation allows the player access to
arenas not along the chosen path, largely defeating the purpose of there
being a choice in the first place. Although teleport scrolls are
limited, they are plentiful and even a single teleport has a high chance
of granting the player access to several new arenas. Unlimited sources
of teleport like the ring of teleportation and the teleportitis mutation
will always let a player loot all areas.

This commit disallows teleports in Gauntlets in the same way as is done
for Sprint. Teleportation scrolls are marked as useless while in the
portal, and attempting to use one aborts with a message. Any passive
form of teleport such as that from the ring, teleportitis, or distortion
is prevented.