Remove additional exits from a Gauntlet map
Remove the intermediate exits from gammafunk_gauntlet_branching. It's
better to have maps be more consistent about the use of exits in terms
of not providing many of them. A future map will require the player to
exit through the minotaur if they don't wish to continue through the
arenas.

We might end up going the route of offering the minotaur vault exit as
the only one for all maps, or we might go the route of this map and have
all maps present an initial exit. This could depend on how we decide to
vary the final fight. If there is no initial exit, it's important that
the player know what strong monster they'll be forced to fight before
entering the portal.