Disable met targets after quaffing a potion of experience (11352)
The potion of experience skill menu backed up the player's current skill
training state to allow potion-specific skill selection. Upon restore, the
backup would ignore any met targets and resume training a skill even if the
skill target had been met.

The skill_state object now checks to see if its cached target has been met
on restore, and does not alter the current state if its training information
appears to be out of date.

Finally, a check is added to the potion of experience effect to prompt for
new skilling if targets are met after quaffing the potion.

Closes #886, using a different approach.