Save compat for 04293e54a7a0
This places an exit under the player (overwriting any features that are
there, so it could overwrite a portal). It is possible for the player to
ignore the exit under some circumstances and get trapped again by taking
a portal, but at that point it's on them. Since the bug wasn't
guaranteed to trigger in this map, this may also provide an early exit
for some players.