Allow Gauntlet maps to customize extra loot
This commit adds Lua support allowing Gauntlet maps to change what loot
their arenas place. This customization only applies to the three extra
item locations within each arena, with the arena's first item remaining
guaranteed. This functionality will be used in a subsequent commit to
have a Gauntlet map place a set of four arenas with the same average
amount of loot as a set of two arenas in another path.