Correctly handle LOS for dispersal traps
Dispersal traps are supposed to affect actors that can see the cell of
the trap without any translocation-blocking wall in the way (e.g.
transparent rock or stone). This condition was correctly applied to
monsters, but not to players if the trap was triggered by a monster.
This commit applies the condition to player, preventing monsters on the
side of transparent rock/stone walls from blinking the player with
dispersal traps.