Fix cross-compiling + libpng
When cross-compiling, the contrib libraries are not useful for the
tilegen tool, which needs to be run as part of the build process. This
prevents it from trying to use contribs; cross-compiling will therefore
fail if it can't find libpng via pkg-config (this was the status quo for
all builds before 409694ba0a8762, so it doesn't introduce any new
requirements).