Re-add Repel Missiles, rework canceling permabuffs
Re-adds Repel Missiles to players and to its original spellbooks. Also
(because I got lazy and didn't split up commits), reworked casting a
permabuff to cancel it to make it so it doesn't cost MP to do so. Could
not apply this to Transmutations because it works differently, but they
have an End Transformation ability so that's fine.