Give shining eyes Polymorph instead of Malmutate
Fits the theme of slimes and form changing while getting rid of the
Malmutate devil. Also upped the spellpower per HD of Polymorph for
monster casting, which hopefully shouldn't matter as only shining eyes
can cast it normally. Might need to tweak down the modifier a bit if it
is too extreme, but wanted to make sure they were still a threat.