A ghost vault with berserking beasts
A transporter vault for D:7-15, Lair, Orc, non-Slime Lair branches, Elf,
Depths, and Zot. In gammafunk_ghost_berserking_beasts, you encounter the
ghost of a player who stumbled into a cave of beasts with moths of wrath
lurking about. The monsters are animal-intelligence things from the
branch of placement that can be berserked. Monster count and loot are
progressive in the usual ghost vault fashion. An extra piece of
auxiliary armor or jewellery that upgrades in quality by depth is placed
under the ghost.

The layout is the same used in another ghost vault for now. I'll rework
it to be something more uniquely cave-like in a future commit.