Remove some ammo quantities from DES monster specs
When giving ammo to a monster as part of its monster spec, the monster
will be given a reasonable quantity for that ammo type if quantity isn't
specified. The monster gear code doesn't correctly apply
monster-specific quantity bonuses when doing this, which is something
that should be fixed. However those aren't relevant for the instances
changed by this commit.