Ghosts remember their slayers
As far as practical, ghosts know who killed them, with a view to ghosts
eventually seeking revenge. A few wrinkles:

We hate the summoner, not the summon. (It will be unsatisfying for the
player if a nearly-dead summoner pops out a crimson imp to have a ghost
arrive to eye up the crimson imp for the two seconds before the summoner
is polished off. It also avoids crazy stuff like ghosts going after
battlespheres or something.)

Some code has been added to remember the type of the monster that
sticky-flamed or poisoned you, even if it dies in the interim. It is
difficult to fix the case where a swamp dragon (etc) breathes a poison
cloud and it is poisoning from the cloud that kills you after the dragon
dies (but it does work if you die of the cloud while the dragon is
alive...)

There is a giant can of worms left unopened surrounding monster-assisted
suicide; there are some fairly clear cases (killed by damage mirrored by
a death knight), but then, killed by a miscast (single-target spell),
killed by a fireball (dropped on self and one monster... self and more
than one monster), poisoned by own meph cloud - down to some clearly
indeterminate ones (fired an IOOD at >1 monster, blinked into it). If
this idea takes off, I might let out some of the less hairy worms, but
now I'm going for the easy option - do nothing.