Change Malmutate spell effect
From mutating the player to deteriorating all the players stats by
1+(1d4, average of 2 rolls). This will only affect Orbs of Fire as they
are the only monster left that has the Malmutate spell and I'm too lazy
to change the name in the source code for everything. This only affects
the spell version of Malmutate, not any other version that may exist
(i.e. Mnoleg melee, spell miscasts, etc).

Also left a note for myself to look into at some point.