Don't block the entrance to Tomb:2 with a statue (SteadysEmpire)
The original Tomb:1 center vault sometimes blocks access to the hatch to
Tomb:2 with a sarcophagus (really a statue). This can create situations
where the player is stranded in Tomb if a player without a way to
destroy statues is able to enter Tomb:2 another way, for instance
formicid shafting or an extremely lucky teleport. This vault predates
the existence of formicids, a species who's combination of stasis and
self-shafting makes this problem much more likely.

In any case, this is not an interesting puzzle, so remove the chance to
place this statue, but still allow this square to sometimes have a
statue monster or trap as before.

(cherry picked from commit 5c428e77e6dfd5b5988f435fd6efb04d1e61d8dc)