Ghosts seek vengeance
Ghosts now emerge from limbo as neutral monsters, seeking vengeance
on their killers.

How this is meant to work, since I guess here is as good a place
to write it all down as any:

There's a 1 in 2 chance of loading a ghost, not 1 in 3. The
ghost is immediately placed in limbo.

Ghosts may be saved anywhere but D:1-2, the Temple, and
Labyrinths. (Why Labs? Ghosts there give a high probability of a
"free pass" on the minotaur.)

Whenever a monster is created, we check to see if there is a
ghost in limbo which has not yet been assigned a target (an
"original foe") and which was slain by this type of monster. If
so, we make a note that this monster is "hated by the ghost". We
also make such a note 1/27 of the time if the ghost has no slayer
(which may happen, eg, if the player suicides by blinking into
their own orb of destruction.) Summoned monsters are not
eligible for this check.

When a hated monster acts, we check to see if there is a ghost
in limbo which either has no original foe or has this specific
hated monster as its original foe. The hated monster has to be
aware of you (not wandering), not petrifying, petrified, asleep,
or petrified, and visible to you. If it is below maximum
hitpoints, there is a chance of nothing happening, rising from
0% at full HP to 100% at 50% HP. Otherwise we look for a spot to
place the ghost.

Placing a ghost looks for an empty spot which is either the same
distance from the monster as you or one space closer to the
monster. (This prevents the ghost, eg, popping into being
trapping you between it and the monster). This spot must be
visible to you, but have an empty spot next to it which isn't
visible to you; ghosts can still come out of corners of rooms
with no exits, but they appear to have become corporeal just out
of your field of view then moved into LOS.

When a monster targetted by a ghost dies, the ghost is completely
removed immediately. A ghost which is destroyed in play is also
completely removed; it failed to get its vengeance.

Ghosts are hard to distract from their original foes. Every time
a ghost acts, there is a 1 in 9 chance it retargets itself on
its original foe. (This will not cause a ghost that you have
rendered hostile to revert to neutrality, but it will still tend
to favour its original foe.)

If a ghost is targetting its original foe after this check, and
finds it is more than twice-LOS range from its original foe, it
returns to limbo. If it finds its original foe is missing from
the level (eg shafted), it returns to limbo (and the other
mechanics imply you will not see it again).

Whenever you leave a level, any ghost on it is returned to
limbo.

Whenever you leave an ephemeral level, any ghost on it in limbo
are resaved as bones files.

When the game is over, any ghosts in limbo on any level are
resaved as bones files.

In both these cases, a ghost is not saved if it has been loaded from
bones but not succeeded in appearing in-game on 9 consecutive occasions.
This is intended to prevent the bones file directory gradually filling
up with ghosts it is extremely hard to use, like D:3 panlord deaths. The
value of '9' is large because, say, a D:3 Sigmund ghost has a plausible
target but might need a few goes to be loaded into a game where Sigmund
is on D:3.