Changes to ghost appearance
The chance of loading a ghost is now not a flat 1 in 2, but increases
from a 1/3 chance if there is only one bones file available to nearly
certain if there are many.

The number of unsuccessful attempts to appear in a game a ghost is
allowed also varies with this number, so that hard-to-load ghosts don't
clutter up D:3 forever but when there are few files available they can
hang about.

Ghosts can now be found on D:2.

The chance for a ghost to appear is now slightly based on tension, in
the hope that this will make it more likely they show up an interesting
moments (but, I hope, not make them too much less likely to appear at all.)