Boolean vampire undead state
Part 3 of a vampire revision.

Add a new player variable to track whether a vampire is alive or
bloodless. Use this in place of the non-functional hunger checks.

Vampires now transition between Alive and Bloodless via abilities,
Exsanguinate and Revivify. These are interruptable delays of length 5,
with no MP cost, to make them tactically difficult to use.

The abilities are implemented separately so that Exsanguinate can
splatter some blood; it might also be desirable to allow Revivify to be
done without a delay if a corpse is available.