Bloodless Vampires regenerate when monsters are not in LOS
Having Bloodless vampires not regenerate at all created either a tedium
headache for the player or a technical headache automating the switch.
After some testing and discussion it's clear that a better approach will
be to inhibit regen for bloodless vampires when monsters are in LOS,
coupled with some strategic costs for switching form.