Increase costs for Vampire abilities
Bat Form is really strong, in part because of its spamability and low MP
cost. To rein it in, this commit adds a random stat drain of
4+binomial(5, 1/2) to the form. The specific number might be too severe;
the choice of stat drain over skill drain is largely borne of a desire
to avoid more uses of Drain. A secondary reason for the choice is that
mutating the undead causes stat damage, so this is a flavourful choice.

To make the choice of Vampire form strategic, instead of costly in
tactical situations but otherwise trivial, revivification gains a
permanent HP cost, similar to Borg's (but less, a flat 2%). A permanent
(instead of XP gated temporaray) cost makes the choice of blood state
strategic (ultimately pressuring a Vampire to pick one and stick to it)
over the scale of the entire game. The permanent MP cost is flavourful
as well. If this is not severe enough, the % can be increased; otherwise
if a permanent cost is undesirable an XP gated cost can be used
instead.

Exsanguination keeps its delay cost, but doesn't gain any other costs.
The cost is paid only to reverse the process.