Rework faded altars
Previously, worshipping at one of these altars would give you a random
god and 20 piety (one and a bit *s).

Change the altar to give a selection of three random gods, and the same
bonuses as Monk on worship.

XXX TODO: The random gods are not persisted, so you can scum the altar
until you get an optimal set of gods to worship.