Regeneration is now a permabuff
Regeneration presented more of a challenge in that at low spellpower,
it can easily be impossible to cast it on cooldown. I expected at some
point to have problems where permabuff maintenance exceeds MP
regeneration, but not so soon! It works on an alternative scheme where
when MP are full, a separate reservoir of MP used for regeneration
alone is topped up.

Once it is dry, MP regeneration is diverted to refill it at a rate
proportionate to MP/max-MP. Improved spellpower reduces the charge on
the reservoir for each aut of regeneration.

Also fix Song of Slaying using Infusion dur for miscasts (whoops),
decide that brainlessness blocks permabuffs like -Cast.