Second permabuff infrastructure commit
Beware if you are taking these commits. Some bugfixen are in the
commits that make spells into permabuffs. Sorry.

This improves the feedback on why permabuffs aren't working,
dispenses with the whole idea of docking piety intermittently
- having a permabuff your god hates is an edge case and we can
just have it not work.

With any luck adding all these functions for whether a permabuff
is working will pay off in the long run.