Permabuff rework - ongoing MP and hunger costs
This adds ongoing MP regen and hunger costs to permabuffs, such that if
you use one once you will take 1/8 of the normal casting and hunger cost,
but as these durations merge into each other you will have the same costs
as recasting it on cooldown in vanilla. PProj like Regen has its own
special mechanism for MP costs because it would often not be possible
to cast it on cooldown in vanilla.