DMsl as permabuff and a raft of other permabuff changes
DMsl's nominal duration is twice what it was in vanilla crawl
because it can also fizzle when attacked. In vanilla, it went from
having a duration to fizzling when attacked. Hence, making players
pay the full equivalent recast cost would be excessive.

When permabuffs are up, don't show the spell as useless. Have spell
descriptions reflect the fact that they are permabuffs, and that
they could be cancelled.

Nerf Shroud (sorry); miscast checks on being meleed, effectiveness
goes from 8 HP to 13 not 10-15.

Overall Charms debt, and warning light for this. It's still pretty
hard to trigger.

Also some more of the amulet of faith change. Sorry!

We do the whole charging for permabuffs dance even if you have no
MP regen. This will build up Charms debt. It's not like DD really
ever wears guardian spirit anyway.

Change the message in @ when a permabuff is suppressed and turned
off.