Permabuff rework #3 or so?
We refactor the permabuff failure checks, since an increasing amount
of code was duplicated in an increasing number of places.

MP regeneration is penalised for spell failure chance, since in vanilla
you'd have to recast it more than once to make it work. Regen reserve
changes take a similar penalty.

If you fail a miscast check, you have to pass one to get the spell up
again. Before this, you could cast the permabuff, don heavy armour, and
enjoy a nominal_duration between miscasts. If the miscast check is above
100%, you just give up.

We don't warn about the danger of recasting the previous spell if it
would cancel a permabuff.

In the 'z' menu, make it clear if we are dispelling permabuffs.