Permabuff tracking fixes
Make the statement that it's safe to call permabuff_track twice in a
row actually true.

Also, somewhat account for you.time_taken in permabuff_fail_track. It's
still the case that (say) being hit many times in a turn increases the
chance of a Shroud failure; I'm OK with that. But it's not the case that
(say) you should take long actions while regenerating to decrease the
chance of a miscast