Remove piety decay
Piety decay has been around for a *long* time. As far as I can tell it
was added in Crawl 2.80 (crawl/crawl-ancient@bd5d840) which also adds
gods in general. There's no code comments in that file about it, so as
far as I know the original design rationale is lost.

In modern DCSS, piety decay has a few functions, principally as a clock
to encourage progression and a way to balance gods.

Piety decays randomly, and very infrequently. On average handle_god_time
is called every 200 aut, and piety is lost on 1/17th or 1/35th of these
calls (depending on the god). Over 80k turns, this translates to losing
230/115 piety. It's hard to get a sense of exactly how much piety is
gained in a typical game, but assuming 2.5k monsters killed worshipping
Trog and 0.7piety per kill (piety per kill depends on your XL and if the
monster counts as a "wizard"), you gain around 1750 piety in this time.
So piety decay accounts for around 15% of piety in this case.

The above paragraph is very rough and ignores a whole bunch of factors,
but does give a ballpark figure of the magnitude of this change. A
future commit will slow down piety gain to balance this removal.