Ring of Flames as a permabuff
You get charged either when you cast a fire spell or when a monster next
to you gets toasted. This isn't quite right - the clouds can be non-
adjacent (but there's no distinguishing them from clouds from eg a wand
of clouds), and there's no charge for blocking a bolt of cold, but this
is about the best I can come up with.