new background: Spellsword
A new background in the "Warrior-mage" section, which currently has only
vanilla backgrounds. There's not a lot of space left in this area
without stepping on the toes of existing backgrounds, or creating new
warrior-magey spells.

Spellswords start with skill in armour and conjurations, and a lot of
extra strength and intelligence. They also get a plain weapon and the
book of conjurations.

The goal is to encourage the "heavy armour caster" archetype that
everyone loves to suggest. This takes inspiration from the bcrawl Reaver
(which is at least partly inspired by the old vanilla Reaver), but it's
sufficiently different I gave it a new name.