Positional magic NNN: Nerf conjure flame (minmay)
Smite targeted, renewable flame clouds are extremely strong.

This changes Conjure flame to create the cloud on the player. The player
gets immunity immediately after its creation until their next action;
after that they burn like everyone else.

Conjure flame is still very strong; and in this version can be used to
create a gap by an escaping player, as well as many of its old uses.

It also did not work well as a monster spell; the only monsters that
still cast it were player ghosts, so it's been removed from them and is
now player only.