Positional magic NNN: New Fire spell: Starburst
Starburst provides a L6 damage spell fitting into a similar place as the
old Bolt of Fire spell. Where Bolt of Fire was targeted, Starburst is
un-targeted.

Instead, Starburst fires four range 5 (six under Vehumet) bolts of fire
in the cardinal directions. The exact targets are fuzzed slightly (in a
symmetric fashion). The result is that a beam path is only guaranteed to range
2. Range 6 with fuzzing increases the guaranteed range to 3. The
guaranteed range can also be increased by using the spell in corridors.

[ TODO: Check how this works with god conduct triggers; add friendly
  fire checks. Create a tile. ]