Positional magic NNN: Modify inner flame
Inner flame is a very positional spell; unfortunately it's somewhat
impractical to use, explosions come only when the monster dies. To open
up new possibilities, this commit adds an on-damage effect to the
enchantment. Inner Flamed monsters have a chance of creating a
short-lived flame cloud on their own cell when they take damage.