Summary: Remove most sources of flight and create ice and obsidian bridges to replace them.
Allow Bolt of Cold, Iceblast and Glaciate to make temporary ice (and obsidian) bridges.
Additionally allow Phials of Floods to make obsidian bridges.
Duration of ice bridges depends on spelldamage (rolled randomly per tile).
If the duration of an ice bridge is less than 5 turns, a warning message prints (using the DANGER channel if you're standing on it).
The tile also changes when duration is less than 5 on tiles.
Fire attacks over an ice bridge (but not an obsidian bridge) decrease the duration based on spellpower.
Actors inside water when an ice bridge would form prevent ice from placing on the tile they are in, but get the Frozen debuff.
Actors inside lava when obsidian forms do not prevent the bridge from forming, but get a new debuff that makes them completely immobile.

Remove Rings of Flight, Potions of Flight and Flight armour brand. (Racial flight, tornado flight and Dragon form flight are still here).
Remove references to emergency flight (which has been gone for awhile now).

When a vault would always place a Ring or Potion of flight, it now places a scroll of random uselessness. Praise Xom.
If a vault had a chance to place one but didn't always; that is just removed.

[Might be temporary] Boots of Strength replace boots of flying. (If I make a new brand for boots I'll get rid of that, but it's fine for now).