Revise the robe of Clouds
The old robe of clouds was basically a +3 {rElec +Fly} robe; it's
gimmick of spawning mist and rain was almost always flavour.

The new robe is now +3 {+Thunder rCloud rElec}, losing the mist gimmick
in favor of evokable radius 2 storm clouds. The evocation costs 5 MP and
has the same difficulty rating as evokable invis.

[BcadrenCrawl] Old trailing clouds effect maintained and buffed.