New species: Silent Spectre
Based on the species from Bcadren's fork, but simplified to the point of
absurdity. No magic, no invocations, no scrolls. The only thing between
you and taking on Zot:5 with a +0 cursed dagger is a couple of yes_or_no
prompts I added, and taking care not to kill mummies while having
anything equipped lol.