Add US_GHOST for Silent Spectres and Mirror Eidola
There were some issues with both these species which required this:
* Mirror Eidola could chunk heal
* Silent Spectres couldn't drink

Ghosts are a new undead type which:
* Get undead resistances (rPois∞, rTorment, rNeg+++, rC+)
* Doesn't hunger
* Can't mutate
* Can quaff
* Can transform

Currently this undead type is not exposed to players on the game
modifier screen because it has no real drawbacks. It might be nice to
add some and then expose it...

We also add a new species flag to indicate a species has no feet for
boots. It could also be used for Argon (and possibly Astral & Mirror
Eidolon), but that work isn't done here.

This tweaks balance a little. Ghosts can now transform (with spectral
tiles for hydra/dragon).

Fixes #41